requirements.txt
opencv-python-headless==4.7.0.68
mediapipe==0.9.0.1
scikit-learn==1.2.0
streamlit==1.26.0
streamlit_lottie==0.0.3
streamlit_webrtc
pickle-mixin
tensorflow
numpy
gtts
playsound
pydub
Pillow
imutils
plotly